[RA]Added more explicit type checking for MessageDigestFactory and moved the Factories to new package from utils
