+++ src/test/java/com/rakole/tinyurl/util/MessageDigestFactoryTest.java
@@ -1,6 +1,7 @@
 package com.rakole.tinyurl.util;
 
 import com.rakole.tinyurl.enums.MessageDigestType;
+import com.rakole.tinyurl.factories.MessageDigestFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,6 +9,7 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class MessageDigestFactoryTest {
 
@@ -23,7 +25,11 @@ class MessageDigestFactoryTest {
         } catch (NoSuchAlgorithmException e) {
             e.printStackTrace();
         }
+    }
 
-
+    @Test
+    @DisplayName("should throw exception when wrong algorithm type sent")
+    void testWhenWrongAlgorithmTypeSent() {
+        assertThrows(NoSuchAlgorithmException.class, () -> MessageDigestFactory.getMessageDigest(null));
     }
 }
